# openapi_dump.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraderCopilot Backend",
    "version": "2.0.1"
  },
  "paths": {
...
    "/analyze/lite": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Lite",
        "description": "Wrapper seguro para capturar errores 500 y mostrarlos en el frontend.\nEnforces Token Gating.",
        "operationId": "analyze_lite_analyze_lite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/analyze/pro": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Pro",
        "operationId": "analyze_pro_analyze_pro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/logs/recent": {
      "get": {
        "tags": [
          "Logs",
          "logs"
        ],
        "summary": "Get Recent Logs",
        "description": "Obtiene las se\u00f1ales m\u00e1s recientes (Paginadas).",
        "operationId": "get_recent_logs_logs_recent_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "saved_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Saved Only"
            }
          },
          {
            "name": "include_system",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include System"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntry"
                  },
                  "title": "Response Get Recent Logs Logs Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/{mode}/{token}": {
      "get": {
        "tags": [
          "Logs",
          "logs"
        ],
        "summary": "Get Logs By Token",
        "description": "Endpoint de compatibilidad para fetchLogs.\nDevuelve logs filtrados por modo y token (User + System).",
        "operationId": "get_logs_by_token_logs__mode___token__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mode"
            }
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/track": {
      "post": {
        "tags": [
          "Logs",
          "logs"
        ],
        "summary": "Create Tracked Signal",
        "description": "Guarda expl\u00edcitamente una se\u00f1al enviada por el frontend (Track).\nUsed when user clicks 'Track' on a transient signal from Analyst.",
        "operationId": "create_tracked_signal_logs_track_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Signal Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/logs/{signal_id}/toggle_save": {
      "post": {
        "tags": [
          "Logs",
          "logs"
        ],
        "summary": "Toggle Save Signal",
        "description": "Toggle the 'is_saved' status of a signal.\nOnly allows modifying own signals or system signals (by cloning? No, just tracking).\nRefined logic: If signal is System (user_id=None), we can't edit it\ndirectly if we want to share it.\nBut for now, assuming simple ownership or shared mutability for System\nsignals is NOT safe.\n\nActually, system signals are shared. If User A tracks it, does it track for User B?\nIf 'is_saved' is on the Signal row, yes. This is a schema flaw for shared system signals.\nHowever, LITE signals are often generated on the fly or by scheduler.\n\nFOR NOW: We assume signals in 'logs/recent' are mostly user-triggered or isolated.\nIf it's a System signal, we might need to clone it to user space to \"Track\" it individually.\n\nStrategy:\nIf user owns signal -> Toggle is_saved.\nIf signal is System (user_id=None) -> CLONE it to user_id=current_user.id\nand set is_saved=1.",
        "operationId": "toggle_save_signal_logs__signal_id__toggle_save_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LiteReq": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Token base: eth|btc|sol|xau"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "description": "Timeframe, e.g. 30m, 1h, 4h"
          },
          "mode": {
            "type": "string",
            "const": "LITE",
            "title": "Mode",
            "default": "LITE"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "timeframe"
        ],
        "title": "LiteReq"
      },
      "LogEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token",
            "default": "UNKNOWN"
          },
          "timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframe",
            "default": "1h"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction",
            "default": "neutral"
          },
          "entry": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry",
            "default": 0.0
          },
          "tp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tp",
            "default": 0.0
          },
          "sl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sl",
            "default": 0.0
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "default": 0.0
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "default": "System"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode",
            "default": "LITE"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "OPEN"
          },
          "pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "exit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Price"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "LogEntry"
      },
...
      "ProReq": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Token base: eth|btc|sol|xau"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "description": "Timeframe, e.g. 30m, 1h, 4h"
          },
          "user_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Message",
            "description": "Mensaje opcional del usuario para contextualizar el an\u00e1lisis PRO"
          }
        },
        "type": "object",
        "required": [
          "token",
          "timeframe"
        ],
        "title": "ProReq"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}